Struggled with alinging the text to center
This error came from the thought process that the assignment
wanted us to only use flexbox and alignment css to align the
items, not taking into account that it wanted us to incorporate
all css that we have learned at this point. The solution to align
all the text was to simply put text-align: center; in the body.